THnSparse, FONLL, AP & tpcPid, HFFitter random seed, Correlated background (#9)
* [qa2] init yield_lifetime_qa_thn

* [qa2] dev yield_lifetime_qa_thn

* [QA] bugfix: do not forget rapidity and datatype cut

* [QA] init dataVsFonll

* add ptLb for Lb spectrum pure from Pythia

* [QA] add ArmenterosPodolanskiCascade (to be re-arranged folders)

* [HFInvMassFitter] random seed fitter

* [qa2, CMakeLists] init multifit_qa, prepend exes with qa2-

* [HFInvMassFitter] bugfix random gaus value

* [qa2] develop multifit_qa; otf

* [o2] copyStandardDerivedData: foresee presence of only AnalysisResults.root or only AO2D.root

* [QA] tpc_qa: foresee different names of the tree

* [o2] tpc configs as in Aug-Sep

* [o2] adjust copyStandardDerivedData for separate processing AO2D.root and AnalysisResults.root (to be tested)

* [bdt] T,M KF->default; smart reading input (exceptions and not too much)

* [Other] re-arrange AP drawers

* [QA] init corrBkgLc (borrow from M.F.)

* [QA] first working application of corrBkgLc

* [QA] vectorize corrBkgLc

* [QA] finish vectorize corrBkgLc

* [QA] add Fit_Lc_flarefly; generalize pt in corrBkgLc

* [QA] corrBkgLc: enable canvasing w/o histos filling (for merged histos)

* [qa2] yield_lifetime_qa_thn: bugfix merge histos when no weights

* [o2] copyStandardDerivedData: smart copy - checks if fully copied

* [QA] corrBkgLc: run from filelist, save meta info

* [qa2] enable filelists; otf mass_bdt_qa_thn

* [cmake] install symlinks on exes

* [QA] corrBkgLc: lifetime-wise

* [QA] tpc_qa: add n sigma tof

* [QA, qa2] tpc_qa: add nSigmaTof histos and cuts

* [qa2] multifit_qa: all vars; complex_vs_bdt_pdfer: skip missing vars; otf

* [o2] add checkForNonRepeating

* [QA] corrBkgLc: typofix DsToPiKK[Pi0] decay formula; drawing improvements

* [qa2] init mass_fit2, otf ShapeFitter

+++ qa2/exe_based/HelperPlot.hpp
@@ -36,7 +36,7 @@ std::pair<double, double> GetMinMaxBinWithError(const TH1* h);
 
 void SetLineDrawParameters(std::vector<TF1*> fs, int lineWidth = 1, int lineStyle = 7, Color_t lineColor = kBlack);
 
-TF1* HorizontalLine4Graph(float level, TGraph* graph);
+TF1* HorizontalLine4Graph(double level, TGraph* graph);
 
 TPaveText* AddOneLineText(const std::string& text, const std::array<float, 4>& xy, const std::string& option="brNDC", float size=0.03);
